Add files via upload
In this project I have involved React js which is javascript library to do events handling in the Todo list and using "useState" React Hook to handle the state in the react Functional Component which is App.js, and also I have attached the final output screenshot of the todolist project. I hope this is the better Stater for learning React js as a beginner.